pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "graphiq"
version = "0.1.0"
description = "LLM-native Graph Exploration System for SAP O2C data"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "asyncpg>=0.29.0",
    "neo4j>=5.20.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "alembic>=1.13.0",
    "structlog>=24.2.0",
    "rapidfuzz>=3.9.0",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.8.0",
    "groq>=0.11.0",
    "psycopg2-binary>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27.0",
    "mypy>=1.10.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app"]
exclude = ["data", "frontend", "migrations"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
strict = true
ignore_missing_imports = true